Refactored utils and changed class to be the only node
