+++ datamodel-ui/src/modules/graph/node.tsx
@@ -9,27 +9,39 @@ import {
   setHovered,
   setSelected,
 } from '@app/common/components/model/model.slice';
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Handle, Position } from 'reactflow';
 import {
   IconChevronDown,
   IconChevronUp,
+  IconMenu,
   IconOptionsVertical,
+  IconSwapVertical,
 } from 'suomifi-ui-components';
-import { ClassNodeDiv, CollapseButton, OptionsButton } from './node.styles';
+import {
+  ClassNodeDiv,
+  CollapseButton,
+  OptionsButton,
+  Resource,
+} from './node.styles';
 import { useStoreDispatch } from '@app/store';
 import { getLanguageVersion } from '@app/common/utils/get-language-version';
 import { useTranslation } from 'next-i18next';
+import { ResourceType } from '@app/common/interfaces/resource-type.interface';
 
 interface ClassNodeProps {
   id: string;
   data: {
     identifier: string;
     label: { [key: string]: string };
+    resources?: {
+      label: { [key: string]: string };
+      identifier: string;
+      type: ResourceType.ASSOCIATION | ResourceType.ATTRIBUTE;
+    }[];
     resourceType?: 'association' | 'attribute';
     applicationProfile?: boolean;
-    toggleResourceVisibility: (value: boolean) => void;
   };
   selected: boolean;
 }
@@ -44,13 +56,6 @@ export default function ClassNode({ id, data }: ClassNodeProps) {
   const [showAttributes, setShowAttributes] = useState(true);
   const [hover, setHover] = useState(false);
 
-  const handleToggleResourceVisibility = useCallback(
-    (value: boolean) => {
-      data.toggleResourceVisibility(value);
-    },
-    [data]
-  );
-
   const handleTitleClick = () => {
     if (globalSelected.id !== id) {
       dispatch(setSelected(id, 'classes'));
@@ -59,7 +64,6 @@ export default function ClassNode({ id, data }: ClassNodeProps) {
 
   const handleShowAttributesClick = () => {
     setShowAttributes(!showAttributes);
-    data.toggleResourceVisibility(showAttributes);
   };
 
   const handleHover = (hover: boolean) => {
@@ -73,8 +77,7 @@ export default function ClassNode({ id, data }: ClassNodeProps) {
 
   useEffect(() => {
     setShowAttributes(globalShowAttributes);
-    handleToggleResourceVisibility(!globalShowAttributes);
-  }, [globalShowAttributes, handleToggleResourceVisibility]);
+  }, [globalShowAttributes]);
 
   return (
     <ClassNodeDiv
@@ -109,6 +112,34 @@ export default function ClassNode({ id, data }: ClassNodeProps) {
           </CollapseButton>
         )}
       </div>
+
+      {showAttributes &&
+        data.resources &&
+        data.resources.map((r) => (
+          <Resource
+            key={`${id}-child-${r.identifier}`}
+            className="node-resource"
+            onClick={() => dispatch(setSelected(r.identifier, 'attributes'))}
+            $highlight={
+              globalSelected.type === 'attributes' &&
+              globalSelected.id !== '' &&
+              globalSelected.id === r.identifier
+            }
+          >
+            {data.applicationProfile &&
+              (r.type === ResourceType.ASSOCIATION ? (
+                <IconSwapVertical />
+              ) : (
+                <IconMenu />
+              ))}
+
+            {getLanguageVersion({
+              data: r.label,
+              lang: displayLang !== i18n.language ? displayLang : i18n.language,
+              appendLocale: true,
+            })}
+          </Resource>
+        ))}
     </ClassNodeDiv>
   );
 }
